pyproject.toml
[project]
name = "pr-review-dataset"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-auth>=2.48.0",
    "google-cloud-bigquery>=3.40.0",
    "httpx>=0.28.1",
    "tqdm>=4.67.3",
    "aiosqlite>=0.20.0",
    "asyncpg>=0.30.0",
    "python-dotenv>=1.0.0",
    "openai>=1.50.0",
    "pydantic>=2.0.0",
    "streamlit>=1.40.0",
    "plotly>=5.24.0",
    "pandas>=2.0.0",
    "psycopg>=3.3.2",
]
